refactor: rename tx-history icon to history
